renaming rewards page to cycles page
