+++ src/router/index.js
@@ -2,7 +2,7 @@ import Vue from 'vue'
 import Router from 'vue-router'
 import Hello from '@/components/Hello'
 import Config from '@/components/Config'
-import Rewards from '@/components/Rewards'
+import Cycles from '@/components/Cycles'
 import CycleInfo from '@/components/CycleInfo'
 
 Vue.use(Router)
@@ -21,9 +21,9 @@ export default new Router({
       component: Config
     },
     {
-      path: '/rewards',
-      name: 'Rewards',
-      component: Rewards
+      path: '/cycles',
+      name: 'Cycles',
+      component: Cycles
     },
     {
       path: '/cycle_info/:cycle',
